requirements.txt
graphiti-core
openai
python-dotenv
fastapi
uvicorn[standard]
pydantic
pydantic-settings

